device: skip IP configuration phase for external devices
We already avoid committing the IP configuration for external devices
(see commit 60334a2893fe). However, we still start DHCP/IPv6-autoconf
and, especially, we change sysctl values of the device.

To be sure that no action is taken on the device, return early from
the IP configuration phase, as in the method=disabled/ignore case.

https://bugzilla.redhat.com/show_bug.cgi?id=1530288