ifcfg-rh: preserve NULL wifi mode when persisting a connection
The wireless mode property can be unset (NULL), in which case it
assumed to be equivalent to "infrastructure". If we convert an unset
mode to infrastructure, the connection will change on write,
triggering errors like:

 settings-connection[...]: write: successfully updated (ifcfg-rh: persist (null)), connection was modified in the process
 device (wlp4s0): Activation: (wifi) access point 'test1' has security, but secrets are required.
 device (wlp4s0): state change: config -> need-auth (reason 'none', sys-iface-state: 'managed')
 device (wlp4s0): The connection was modified since activation
 device (wlp4s0): state change: need-auth -> failed (reason 'no-secrets', sys-iface-state: 'managed')

To fix this, remove the MODE key when the mode is unset so that the
property is read back exactly as it was. Note that initscripts need
the MODE set, but in most cases there are other issues that make Wi-Fi
connection written by NM not compatible with initscripts.

https://bugzilla.redhat.com/show_bug.cgi?id=1549972